# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BotAluguel Pro API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bots:
    get:
      operationId: listBots
      tags: [bots]
      summary: List user's bots
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of bots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bot"
    post:
      operationId: createBot
      tags: [bots]
      summary: Create a new bot
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBotBody"
      responses:
        "201":
          description: Bot created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"
  /bots/{botId}:
    get:
      operationId: getBot
      tags: [bots]
      summary: Get bot details
      security:
        - bearerAuth: []
      parameters:
        - name: botId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bot details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"
    delete:
      operationId: deleteBot
      tags: [bots]
      summary: Delete a bot
      security:
        - bearerAuth: []
      parameters:
        - name: botId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Bot deleted
...
  /hosted-bots:
    get:
      operationId: listHostedBots
      tags: [hosting]
      summary: List user hosted bots
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of hosted bots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HostedBot"
    post:
      operationId: createHostedBot
      tags: [hosting]
      summary: Upload a Node.js bot (ZIP or GitHub)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                githubUrl:
                  type: string
                file:
                  type: string
                  format: binary
              required:
                - name
      responses:
        "200":
          description: Bot created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HostedBot"
  /hosted-bots/{id}:
    get:
      operationId: getHostedBot
      tags: [hosting]
      summary: Get a hosted bot by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Hosted bot details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HostedBot"
    delete:
      operationId: deleteHostedBot
      tags: [hosting]
      summary: Delete a hosted bot
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bot deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /hosted-bots/{id}/start:
    post:
      operationId: startHostedBot
      tags: [hosting]
      summary: Start a hosted bot process
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bot started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /hosted-bots/{id}/stop:
    post:
      operationId: stopHostedBot
      tags: [hosting]
      summary: Stop a hosted bot process
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bot stopped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /hosted-bots/{id}/restart:
    post:
      operationId: restartHostedBot
      tags: [hosting]
      summary: Restart a hosted bot process
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bot restarted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /hosted-bots/{id}/files:
    get:
      operationId: getHostedBotFiles
      tags: [hosting]
      summary: List files in a hosted bot project
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File list
          content:
            application/json:
              schema:
                type: object
                properties:
                  files:
                    type: array
                    items:
                      type: string
  /hosted-bots/{id}/file:
    get:
      operationId: getHostedBotFile
      tags: [hosting]
      summary: Read a file from a hosted bot project
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: path
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File content
          content:
            application/json:
              schema:
                type: object
                properties:
                  path:
                    type: string
                  content:
                    type: string
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    HostedBot:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        name:
          type: string
        sourceType:
          type: string
        githubUrl:
          type: string
          nullable: true
        mainFile:
          type: string
        status:
          type: string
        isRunning:
          type: boolean
        pid:
          type: number
          nullable: true
        files:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Bot:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        name:
          type: string
        phone:
          type: string
          nullable: true
        status:
          type: string
          enum: [disconnected, connecting, connected, error]
        connectionType:
          type: string
          enum: [qrcode, code]
          nullable: true
        qrCode:
          type: string
          nullable: true
        pairCode:
          type: string
          nullable: true
        totalGroups:
          type: number
        createdAt:
          type: string
          format: date-time
        connectedAt:
          type: string
          nullable: true
          format: date-time
        prefix:
          type: string
        ownerPhone:
          type: string
          nullable: true
      required:
        - id
        - userId
        - name
        - status
        - totalGroups
        - createdAt
        - prefix
    CreateBotBody:
      type: object
      properties:
        name:
          type: string
      required:
        - name
...